Implemented a test suite in test/. So far it has the toy data sets and the motion capture data set from the jupyter notebooks. Running python run_test.py runs SEM on those and reports the segmentation performance. It also generates a notebook with some plots. It is useful enough to do quick-n-dirty sanity checks of any changes in the code.
I decided to make it part of the project for now as I use a lot of the
modules and utils. Also, it might be convenient to reuse some of the
testing methods in the notebooks too to avoid code duplication (I
actually copy-pasted most of them from the notebooks).

Next steps are:
- include coffee shop world
- random test generator (for stress testing)
- summary statistics
- report other metrics, e.g. scene prediction